gammastep: use manual provider
this is due to geoclue service is no longer able to query location using
the Mozilla Location Services as it has been retired. an alternative
would be to use Google Location API, which requires an API key, but I'm
too lazy to set that up. for now hardcoding coords is fine.

reference: https://unix.stackexchange.com/questions/778391/geoclue-failed-to-query-location-not-found