feat(home/desktop/eww): dynamic margin and borders
since in Hyprland we set up window gaps and border radiuses only if
there are multiple windows, it makes sense to try to fit our eww bar to
have a consistent experience alike.

a margin and border radius is added to the topbar widget if there are
multiple windows open. if there's only one window open, then the margin
and border radius is removed so the window is flush to the bar.