feat(hosts/common/upgrade): avoid building locally
an additional check was added to nixos-upgrade system unit which
prevents the system trying to build the config locally. it looks for the
top level derivation on upstream binary caches. (or if it already exists
locally in the store, it skips rebuilding anyway.)

this has the side effect of preventing a broken config to be
downstreamed. as Buildbot successfully builds the new config, it pushes
the results to binary caches. only then will hosts try to rebuild.