chore(hosts/common/auto-upgrade): buildbot check
for some reason, the output hash of each NixOS host config is different
depending on where it was evaluated. because of this, we cannot depend
on the store path being present on caches for checking while auto
upgrading.

to overcome this, we instead check if the CI/CD job is successful for the
latest revision. this in theory should work now.