fix(hosts/gaia/network-entities): handle dangling links
previously, the script only took into account existing links. for
example: if an rndis interface is unplugged, it will no longer exist,
thus the state will be stuck in still being connected.

this change adds handling a predefined set of interfaces. if any of them
is missing, send disconnected state.