fix(hosts/gaia): let auto-upgrade continute with offline substituters
previously, if one of the systems were offline, the upgrade would fail
because of `nix.settings.fallback = false;`. this reverts that setting.

this also allows nix to continute building if one or more substituters
are offline. this should not be a problem, because auto-upgrade is only
run if it is already built by CI/CD, so it would't eat away at system
resources.