fix(hosts/common/certs): persist acme state
this prevents crashing `acme-{...}.service` units after a reboot, when
the host can't reach the acme server yet. we can safely assume, that
certificate files are already present, so the units won't fail trying to
immediately request a new one. this will probably not be the case on
first boot. that case isn't all that important anyway...