feat(hosts/gaia/network): add transient wan config
this trickery allows us to plug in an Android ethernet tether to any LAN
port. i didn't find any source on what ip ranges are used, but it seems
that it consistently has the IP 192.168.51.128/24. so by adding a static
IP to `br-lan` on this subnet (192.168.51.254/24), we can reach this
gateway now from our router, and it can be used as a failover if `wan0`
goes down. cool.

i call this "interface" transient because it's not really an interface
on the router, but just another device on another subnet.