fix(hosts/common/auto-upgrade): select only one change for revision
this fixes the exec-condition not firing on a rare occasion, when
it should.

if a commit was pushed onto multiple branches, then multiple changes
would be returned on the buildbot api. i think it's safe to assume
this is a relatively rare occasion, so we blindly select
the latest change.