feat(hosts/common/auto-upgrade): allow switching outside reboot window
this changes the upstream logic regarding the auto upgrade reboot
window, which prior to this commit, was replicated in the custom
script.

now, the `operation` attribute is respected regardless of
`rebootAllowed`. this means, if there is a reboot window configured
and rebooting is allowed, and if we're outside the reboot window,
`operation` is not overwritten anymore.

this allows the ability for the following example scenario: if we
want to only reboot machines outside working hours for kernel updates,
etc., and also have the ability to push small configuration changes
during the day, it's possible via setting `operation = "switch";` and
of course setting both `rebootWindow` and `rebootAllowed = true;`