chore: nix-colors -> stylix
ripping out the old, and putting in the new styling
framework.

this tries to make no runtime modifications regarding
functionality. goal is to not change look and feel with this
commit, just if necessary.

notable changes:
- GTK theme is now based on adw-gtk3
- xsettingsd doesn't work; GTK apps don't reload if
light/dark theme is toggled (because the theme name doesn't
change, thus xsettingsd skips reloading)
- declaring the color scheme is now nixos-system-wide, not
just home-manager-wide. specialisations for dark and light
still only reside in home-manager for now.