feat(esphome): firmware versioning
this enables us to skip unnecessary OTAs by comparing the
on-device version against the target version.

the firmware version is derived from the git revision of the
system flake running the OTA update services. this is nice,
because if the tree is dirty, then we can skip checking the
on-device version. this allows to prototype esphome configs
easily.