feat(gaia/net): wan failover using health checking
this is a revisited attempt of deef51a.
the difference is that now we use two routing tables: `wan-primary` and
`wan-failover`. the former is active when the primary wan link is deemed
to be healthy. the latter is activated when the primary link doesn't seem
healthy. the heuristic for "health" is a simple ping for now, which
could be expanded to test other means of internet accessibility.

a prerouting firewall rule is used to mark all packets. the value of the
mark is determined by the health checking algorithm outlined above.
each routing table is assigned to a mark. this is what enables us to
change routing on-the-fly.