chore(global,secrets): remove old gpg key
related to 16a7aaf (second step in rotating keys)

this phases out and essentially soft-expires the
pre-rotation GPG key. we don't want to revoke the key, as
it is still considered valid for historic data (eg. signed
git commits), but is it won't be used for new data and
encryption/decryption.