fix(modules/nixos/esphome-updater): use docker for reproducibility
closes #306

see the issue described in the issue above. the issue was
resolved by issuing the docker image provided by the esphome
project for running device updates.

it also has a nice side effect of being able to choose which
esphome framework version we want to use, per device! this
change adds support for that.