fix(modules/esphome-updater): pin esphome version to the one in nixpkgs
previously using the `latest` docker image tag was ahead
of the version found in nixpkgs, which means that Home
Assistant's ESPHome integration cannot connect to devices as
it is older (using `services.home-assistant`).

this change by default pins the docker image to the same
version to the one found in the current nixpkgs branch.